+++ src/main/java/ee/carlrobert/codegpt/completions/llama/LlamaServerMessage.java
@@ -3,5 +3,5 @@ package ee.carlrobert.codegpt.completions.llama;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record LlamaServerMessage(String level, String message) {
+public record LlamaServerMessage(String level, String msg) {
 }