+++ src/main/java/ee/carlrobert/codegpt/completions/llama/LlamaServerStartupParams.java
@@ -3,5 +3,6 @@ package ee.carlrobert.codegpt.completions.llama;
 import java.util.List;
 
 public record LlamaServerStartupParams(String modelPath, int contextLength, int threads, int port,
-                                       List<String> additionalParameters) {
+                                       List<String> additionalRunParameters,
+                                       List<String> additionalBuildParameters) {
 }