+++ src/main/java/ee/carlrobert/codegpt/settings/service/llama/form/LlamaSettingsForm.java
@@ -41,6 +41,7 @@ public class LlamaSettingsForm extends JPanel {
     state.setContextSize(llamaServerPreferencesForm.getContextSize());
     state.setThreads(llamaServerPreferencesForm.getThreads());
     state.setAdditionalParameters(llamaServerPreferencesForm.getAdditionalParameters());
+    state.setAdditionalBuildParameters(llamaServerPreferencesForm.getAdditionalBuildParameters());
 
     var modelPreferencesForm = llamaServerPreferencesForm.getLlamaModelPreferencesForm();
     state.setCustomLlamaModelPath(modelPreferencesForm.getCustomLlamaModelPath());