+++ src/main/kotlin/ee/carlrobert/codegpt/actions/CodeCompletionFeatureToggleActions.kt
@@ -2,11 +2,12 @@ package ee.carlrobert.codegpt.actions
 
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.components.service
 import com.intellij.openapi.project.DumbAwareAction
 import ee.carlrobert.codegpt.settings.GeneralSettings
 import ee.carlrobert.codegpt.settings.service.ServiceType
-import ee.carlrobert.codegpt.settings.service.ServiceType.LLAMA_CPP
-import ee.carlrobert.codegpt.settings.service.ServiceType.OPENAI
+import ee.carlrobert.codegpt.settings.service.ServiceType.*
+import ee.carlrobert.codegpt.settings.service.custom.CustomServiceSettings
 import ee.carlrobert.codegpt.settings.service.llama.LlamaSettings
 import ee.carlrobert.codegpt.settings.service.openai.OpenAISettings
 
@@ -14,12 +15,16 @@ abstract class CodeCompletionFeatureToggleActions(
     private val enableFeatureAction: Boolean
 ) : DumbAwareAction() {
 
+
     override fun actionPerformed(e: AnActionEvent) {
         GeneralSettings.getCurrentState().selectedService
-            .takeIf { it in listOf(OPENAI, LLAMA_CPP) }
+            .takeIf { it in listOf(OPENAI, CUSTOM_OPENAI, LLAMA_CPP) }
             ?.also { selectedService ->
                 if (OPENAI == selectedService) {
                     OpenAISettings.getCurrentState().isCodeCompletionsEnabled = enableFeatureAction
+                } else if (CUSTOM_OPENAI == selectedService) {
+                    service<CustomServiceSettings>().state.codeCompletionSettings.codeCompletionsEnabled =
+                        enableFeatureAction
                 } else {
                     LlamaSettings.getCurrentState().isCodeCompletionsEnabled = enableFeatureAction
                 }
@@ -31,7 +36,7 @@ abstract class CodeCompletionFeatureToggleActions(
         val codeCompletionEnabled = isCodeCompletionsEnabled(selectedService)
         e.presentation.isEnabled = codeCompletionEnabled != enableFeatureAction
         e.presentation.isVisible =
-            e.presentation.isEnabled && listOf(OPENAI, LLAMA_CPP).contains(
+            e.presentation.isEnabled && listOf(OPENAI, CUSTOM_OPENAI, LLAMA_CPP).contains(
                 selectedService
             )
     }
@@ -43,6 +48,7 @@ abstract class CodeCompletionFeatureToggleActions(
     private fun isCodeCompletionsEnabled(serviceType: ServiceType): Boolean {
         return when (serviceType) {
             OPENAI -> OpenAISettings.getCurrentState().isCodeCompletionsEnabled
+            CUSTOM_OPENAI -> service<CustomServiceSettings>().state.codeCompletionSettings.codeCompletionsEnabled
             LLAMA_CPP -> LlamaSettings.getCurrentState().isCodeCompletionsEnabled
             else -> false
         }