+++ src/main/kotlin/ee/carlrobert/codegpt/codecompletions/CodeGPTInlineCompletionProvider.kt
@@ -5,6 +5,7 @@ import com.intellij.codeInsight.inline.completion.InlineCompletionProvider
 import com.intellij.codeInsight.inline.completion.InlineCompletionProviderID
 import com.intellij.codeInsight.inline.completion.InlineCompletionRequest
 import com.intellij.codeInsight.inline.completion.elements.InlineCompletionGrayTextElement
+import com.intellij.notification.NotificationType
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSingleSuggestion
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSuggestionUpdateManager
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSuggestionUpdateManager.UpdateResult
@@ -12,12 +13,17 @@ import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSug
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSuggestionUpdateManager.UpdateResult.Invalidated
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionVariant
 import com.intellij.openapi.application.EDT
-import com.intellij.openapi.diagnostic.Logger
+import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.thisLogger
+import ee.carlrobert.codegpt.CodeGPTKeys
 import ee.carlrobert.codegpt.completions.CompletionRequestService
 import ee.carlrobert.codegpt.settings.GeneralSettings
 import ee.carlrobert.codegpt.settings.service.ServiceType
+import ee.carlrobert.codegpt.settings.service.custom.CustomServiceSettings
 import ee.carlrobert.codegpt.settings.service.llama.LlamaSettings
 import ee.carlrobert.codegpt.settings.service.openai.OpenAISettings
+import ee.carlrobert.codegpt.ui.OverlayUtil.showNotification
+import ee.carlrobert.llm.client.openai.completion.ErrorDetails
 import ee.carlrobert.llm.completion.CompletionEventListener
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.awaitClose
@@ -29,9 +35,8 @@ import okhttp3.sse.EventSource
 import java.util.concurrent.atomic.AtomicReference
 
 class CodeGPTInlineCompletionProvider : InlineCompletionProvider {
-
     companion object {
-        private val LOG = Logger.getInstance(CodeGPTInlineCompletionProvider::class.java)
+        private val logger = thisLogger()
     }
 
     private val currentCall = AtomicReference<EventSource>(null)
@@ -44,7 +49,7 @@ class CodeGPTInlineCompletionProvider : InlineCompletionProvider {
 
     override suspend fun getSuggestion(request: InlineCompletionRequest): InlineCompletionSingleSuggestion {
         if (request.editor.project == null) {
-            LOG.error("Could not find project")
+            logger.error("Could not find project")
             return InlineCompletionSingleSuggestion.build(elements = emptyFlow())
         }
 
@@ -55,12 +60,14 @@ class CodeGPTInlineCompletionProvider : InlineCompletionProvider {
             currentCall.set(
                 CompletionRequestService.getInstance().getCodeCompletionAsync(
                     infillRequest,
-                    CodeCompletionEventListener(infillRequest) {
+                    CodeCompletionEventListener {
+                        val inlineText = it.takeWhile { message -> message != '\n' }.toString()
+                        request.editor.putUserData(CodeGPTKeys.PREVIOUS_INLAY_TEXT, inlineText)
                         launch {
                             try {
-                                trySend(InlineCompletionGrayTextElement(it))
+                                trySend(InlineCompletionGrayTextElement(inlineText))
                             } catch (e: Exception) {
-                                LOG.error("Failed to send inline completion suggestion", e)
+                                logger.error("Failed to send inline completion suggestion", e)
                             }
                         }
                     }
@@ -74,6 +81,7 @@ class CodeGPTInlineCompletionProvider : InlineCompletionProvider {
         val selectedService = GeneralSettings.getCurrentState().selectedService
         val codeCompletionsEnabled = when (selectedService) {
             ServiceType.OPENAI -> OpenAISettings.getCurrentState().isCodeCompletionsEnabled
+            ServiceType.CUSTOM_OPENAI -> service<CustomServiceSettings>().state.codeCompletionSettings.codeCompletionsEnabled
             ServiceType.LLAMA_CPP -> LlamaSettings.getCurrentState().isCodeCompletionsEnabled
             else -> false
         }
@@ -85,24 +93,28 @@ class CodeGPTInlineCompletionProvider : InlineCompletionProvider {
     }
 
     internal class CodeCompletionEventListener(
-        private val requestDetails: InfillRequestDetails,
-        private val completed: (String) -> Unit
+        private val completed: (StringBuilder) -> Unit
     ) : CompletionEventListener<String> {
 
+        override fun onMessage(message: String?, eventSource: EventSource?) {
+            if (message != null && message.contains('\n')) {
+                eventSource?.cancel()
+            }
+        }
+
         override fun onComplete(messageBuilder: StringBuilder) {
-            // TODO: https://youtrack.jetbrains.com/issue/CPP-38312/CLion-crashes-around-every-10-minutes-of-work
-            /*val processedOutput = CodeCompletionParserFactory
-                .getParserForFileExtension(requestDetails.fileExtension)
-                .parse(
-                    requestDetails.prefix,
-                    requestDetails.suffix,
-                    messageBuilder.toString()
-                )*/
-            val output =
-                if (messageBuilder.contains("\n"))
-                    messageBuilder.substring(0, messageBuilder.indexOf("\n"))
-                else messageBuilder.toString()
-            completed(output)
+            completed(messageBuilder)
+        }
+
+        override fun onCancelled(messageBuilder: StringBuilder) {
+            completed(messageBuilder)
+        }
+
+        override fun onError(error: ErrorDetails, ex: Throwable) {
+            if (ex.message == null || (ex.message != null && ex.message != "Canceled")) {
+                showNotification(error.message, NotificationType.ERROR)
+                logger.error(error.message, ex)
+            }
         }
     }