+++ src/main/kotlin/ee/carlrobert/codegpt/util/file/LanguageFileExtensionDetails.kt
@@ -1,0 +1,4 @@
+package ee.carlrobert.codegpt.util.file
+
+@JvmRecord
+data class LanguageFileExtensionDetails(val name: String, val type: String, val extensions: List<String>?)